Delete groups members successly with admin role

+++ app/views.py
@@ -1,7 +1,7 @@
 from audioop import reverse
 from urllib import request
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
+from django.http import HttpResponse, HttpResponseNotAllowed, HttpResponseRedirect, JsonResponse
 import json
 from django.contrib.auth import authenticate, login, logout # thu vien xac thuc
 from django.contrib.auth.models import User
@@ -115,17 +115,29 @@ def addGroup(request):
 
 def deleteGroup(request, group_id):
     group = get_object_or_404(Group, pk=group_id)
-    group.delete()
-    messages.success(request, "Group deleted successfully")
-    return redirect('member-of-trans-team', group_id=group_id)
+    # Kiểm tra xem người dùng hiện tại có trong nhóm và có vai trò admin trong nhóm đó không
+    if Member.objects.filter(auth_user=request.user, group=group, teamrole='admin').exists():
+        if request.method == 'POST':
+            group.delete()
+            messages.success(request, 'Delete group successfully')
+    else:
+        messages.error(request, "You must have admin role in the group to delete this group")
+    return redirect('transteam')
 
 def deleteMember(request, group_id, member_id):
     # Lấy nhóm và thành viên từ cơ sở dữ liệu
     group = get_object_or_404(Group, id=group_id)
     member = get_object_or_404(Member, id=member_id)
-    if member.teamrole == "admin":
-        messages.error(request, "Can't delete an admin member.")
+    if Member.objects.filter(auth_user=request.user, group=member.group).exists():
+            # Kiểm tra xem người dùng hiện tại có vai trò admin trong nhóm đó không
+            if Member.objects.filter(auth_user=request.user, group=member.group, teamrole='admin').exists():
+                if member.teamrole == "admin":
+                    messages.error(request, "Can't delete an admin member.")
+                else:
+                    member.delete()
+                    messages.success(request, "Delete member successfully")
+            else:
+                messages.error(request, "You must have admin role in the group to delete this member")
     else:
-        member.delete()
-        messages.success(request, "Member deleted successfully")
+        messages.error(request, "You must be a member of the group to delete this member")
     return redirect('member-of-trans-team', group_id=group_id)
